test(terminal): drive the unknown-session rejection through the real route (#14961)
AC4 asks for a test that rejects an unknown session_id "driving the real
route". The only unknown-session test called the decorated handler directly,
which cannot observe a routing or dependency-resolution failure -- exactly how
#14998 stayed invisible while this route 500'd for every caller and the
handler-level guard tests stayed green.

Adds a TestClient handshake against a session_id never written to the store,
asserting 1008 on the wire, an accept-spy count of 0, _init_terminal_handler
never awaited, and -- via caplog -- that the unknown-session branch fired and
the ownership branch did not. Both branches exit through the same close(1008),
so without the caplog pair the close code cannot tell them apart and a
regression turning "unknown" into "not the owner" would pass silently. The
non-owner case already pins the mirror direction.

The config store is Redis-backed and fails closed, so an unreachable store
answers every lookup with None and would produce the same 1008 for the wrong
reason. The test reads a real session back out of the fakeredis-backed store
first, which makes the subsequent miss a genuine absence rather than an
outage.

Also splits the structural router/dependency-wiring assertions and their
subprocess route dump into their own module. The host file was 613 lines
against the 600-line guard before this change -- an unflagged violation on the
base -- and the seam taken is the one its own docstring already described: the
structural half needs no fixtures, no fakeredis and no auth stubbing. Both
files are now well inside the limit with no new grandfather entry.

Contrast mutations, restored after each: returning an empty config instead of
None for an unknown session makes the handshake succeed and the test fail on
the missing disconnect; swapping the two log messages leaves 1008, the accept
spy and assert_not_awaited passing and fails on the caplog assertion alone.
